+++ CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+<!-- ## Unreleased -->
 <!-- Add new, unreleased changes here. -->
+
+## [3.6.0] - 2017-11-28
+* Added a new lint option: `filesToIgnore`. We'll never report warnings for any
+  file that matches any of these globs.
+  * `ignoreWarnings` is a confusing phrasing. Added the `warningsToIgnore`
+    spelling which will be used preferentially. This was better than making the
+    options inconsistent, or spelling the new option `ignoreFiles`.
+* Improved error messages when validation of a polymer.json object fails.
+<!-- Add new, unreleased changes here. -->
+
+## [3.5.0] - 2017-11-21
+* Added static methods for constructing a ProjectConfig directly from an unvalidated JSON object, in addition to the methods for reading it from disk.
 
 ## [3.4.0] - 2017-06-21
 * Modified the `bundle` property in project build options to support the subset of `polymer-bundler` options which can be serialized in a JSON file.